Added more tests for the new FunctionPointer/Event implementation
And yes, I know these should be automated, but it turns out that this
kind of test that outputs lots of text is the perfect opportunity to
restart the discussions about an expect-like interface for some of our
tests. So I'll try to have that in place and make this into an
automated test afterwards.
Tested on K64F with both GCC and ARMCC.