Remove 'multiple symbol definitions' linker errors
One of the bind() functions in Event.h is not a template, so it needs to
have its implementation in a separate file, otherwise the linker will
complain about multiple symbol definitions. This is of course in complete
contrast with the templated versions which pretty much _need_ to be
implemented in the header file. Oh C++, you're a riot.