Disable app.h
As decribed in #136 and #118, app.h is redundant and problematic.
This commit replaces the previous code in app.h with a deprecation
warning, instead of removing it completely, to keep compatibility
with code that includes "app.h" explicitly.

Tested by compiling mbed-drivers, minar and core-util with
mbed-k64f-gcc and mbed-k64f-armcc.