Creating the doxygen front page
Doxygen on docs.mbed requires a front page file under the name DOXYGEN_FRONTPAGE.md. This is just a copy of the repo's readme. I will look into transclusion options later on, and if they work I'll replace the contents of this page with a link to the original readme.